bug fixing

+++ basic_nn_MNIST/main.py
@@ -76,7 +76,7 @@ def loss_fn(model, features, config, training):
     mean_classification_loss = tf.reduce_mean(classification_loss) 
 
     accuracy = tf.reduce_mean(
-        tf.where(tf.equal(tf.cast(label, dtype=tf.float32), tf.argmax(model_output, axis=-1)),
+        tf.where(tf.equal(tf.cast(label, dtype=tf.float32), tf.cast(tf.argmax(model_output, axis=-1), dtype=tf.float32)),
                     tf.ones_like(label, dtype=tf.float32),
                     tf.zeros_like(label, dtype=tf.float32)))
